tests/test_configuration_fields.py : update for Annotation
This test started failing with [1].  These annotations are somewhat of
a duplicate of the constraints; one thought is that you could just
parse out the first item of the Annotation - but that might cause more
problems than it solves.

Fix the testing so it works, but just skip this if somehow we have a
lower sphinx to avoid bumping the base requirement.

[1] https://github.com/sphinx-doc/sphinx/commit/159c26715be2f296b57338f37fcc6edd80719a7d